+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### 5.1.2
-* add message to re-activate current, active theme if it has been updated, because it's now broken. :P
+* add `upgrader_source_selection` filter back for correct updating of current, active theme.
+* fix [#293](https://github.com/afragen/github-updater/issues/293) and [#297](https://github.com/afragen/github-updater/issues/297)
 
 #### 5.1.1
 * hotfix to comment out `pre_http_request` filter. Updating of plugin doesn't work. I need to re-think this one.